+++ diskann/src/graph/index.rs
@@ -30,7 +30,7 @@ use super::{
     },
     internal::{BackedgeBuffer, SortedNeighbors, prune},
     search::{
-        Knn,
+        Knn, PagedSearch,
         record::{NoopSearchRecord, SearchRecord, VisitedSearchRecord},
         scratch::{self, PriorityQueueConfiguration, SearchScratch, SearchScratchParams},
     },
@@ -42,7 +42,7 @@ use crate::{
     ANNError, ANNErrorKind, ANNResult,
     error::{ErrorExt, IntoANNResult},
     internal,
-    neighbor::{self, Neighbor, NeighborPriorityQueue, NeighborQueue},
+    neighbor::{self, Neighbor, NeighborQueue},
     provider::{
         Accessor, AsNeighbor, AsNeighborMut, BuildDistanceComputer, BuildQueryComputer,
         DataProvider, Delete, ElementStatus, ExecutionContext, Guard, NeighborAccessor,
@@ -2052,35 +2052,6 @@ where
         }
     }
 
-    /// Filter out start nodes from the best candidates in the scratch.
-    fn filter_search_candidates(
-        &self,
-        start_points: &[DP::InternalId],
-        l_value: usize,
-        best: &mut NeighborPriorityQueue<DP::InternalId>,
-    ) -> impl SendFuture<ANNResult<(Vec<Neighbor<DP::InternalId>>, usize)>> {
-        async move {
-            let mut total = 0usize;
-            let mut candidates = Vec::with_capacity(l_value);
-            for n in best.iter() {
-                total += 1;
-                if !start_points.contains(&n.id) {
-                    candidates.push(n);
-                    if candidates.len() >= l_value {
-                        break;
-                    }
-                }
-            }
-
-            debug_assert!(
-                l_value.min(best.size().saturating_sub(start_points.len())) <= candidates.len(),
-                "Not enough candidates after filtering starting points",
-            );
-
-            Ok((candidates, total))
-        }
-    }
-
     /// Execute a search using the unified search interface.
     ///
     /// This method provides a single entry point for all search types. The `search_params` argument
@@ -2254,7 +2225,7 @@ where
         l_value: usize,
     ) -> impl SendFuture<ANNResult<PagedSearch<'a, DP, S, T>>>
     where
-        S: SearchStrategy<DP, T> + 'static,
+        S: SearchStrategy<DP, T>,
         T: Copy + Send + 'a,
     {
         async move {
@@ -2276,7 +2247,7 @@ where
         init_ids: Option<&'a [DP::InternalId]>,
     ) -> impl SendFuture<ANNResult<PagedSearch<'a, DP, S, T>>>
     where
-        S: SearchStrategy<DP, T> + 'static,
+        S: SearchStrategy<DP, T>,
         T: Copy + Send + 'a,
     {
         async move {
@@ -3055,127 +3026,6 @@ impl InternalSearchStats {
 struct BatchIdMismatch {
     batch_len: usize,
     ids_len: usize,
-}
-
-//////////////////
-// Paged Search //
-//////////////////
-
-/// A paged search handle that owns all search state internally.
-///
-/// Created by [`DiskANNIndex::paged_search`] or
-/// [`DiskANNIndex::paged_search_with_init_ids`]. Each call to
-/// [`next_page`](Self::next_page) resumes the graph search and returns the next page of
-/// nearest-neighbor results. Returns an empty `Vec` when the search is exhausted.
-#[derive(Debug)]
-pub struct PagedSearch<'a, DP: DataProvider, S: SearchStrategy<DP, T>, T> {
-    index: &'a DiskANNIndex<DP>,
-    context: &'a DP::Context,
-    scratch: SearchScratch<DP::InternalId>,
-    computed_result: Vec<Neighbor<DP::InternalId>>,
-    next_result_index: usize,
-    search_param_l: usize,
-    strategy: S,
-    computer: S::QueryComputer,
-    // Note: The `fn` is so we derive `Send` and `Sync` more easily: `fn` is always Send/Sync.
-    _query: std::marker::PhantomData<fn(T)>,
-}
-
-impl<'a, DP, S, T> PagedSearch<'a, DP, S, T>
-where
-    DP: DataProvider,
-    S: SearchStrategy<DP, T>,
-{
-    /// Returns the next page of at most `k` nearest-neighbor results.
-    ///
-    /// Results across pages are non-overlapping and ordered by non-decreasing distance.
-    /// When the search is exhausted, returns an empty `Vec`.
-    pub fn next_page(
-        &mut self,
-        k: usize,
-    ) -> impl SendFuture<ANNResult<Vec<Neighbor<DP::InternalId>>>> {
-        async move {
-            if k > self.search_param_l {
-                return ANNResult::Err(ANNError::log_paged_search_error(
-                    "k should be less than or equal to search_param_l".to_string(),
-                ));
-            }
-            if k == 0 {
-                return ANNResult::Err(ANNError::log_paged_search_error(
-                    "k should be greater than 0".to_string(),
-                ));
-            }
-
-            let mut result = Vec::with_capacity(k);
-
-            // Drain any already-computed results first.
-            let available = self
-                .computed_result
-                .len()
-                .saturating_sub(self.next_result_index);
-            let from_cache = cmp::min(k, available);
-            if from_cache > 0 {
-                result.extend_from_slice(
-                    &self.computed_result
-                        [self.next_result_index..self.next_result_index + from_cache],
-                );
-                self.next_result_index += from_cache;
-
-                if result.len() == k {
-                    return ANNResult::Ok(result);
-                }
-            }
-
-            // Resume graph search to fill the next batch.
-            let start_points = {
-                let mut accessor = self
-                    .strategy
-                    .search_accessor(&self.index.data_provider, self.context)
-                    .into_ann_result()?;
-
-                let start_ids = accessor.starting_points().await?;
-                self.index
-                    .search_internal(
-                        None, // beam_width
-                        &start_ids,
-                        &mut accessor,
-                        &self.computer,
-                        &mut self.scratch,
-                        &mut NoopSearchRecord::new(),
-                    )
-                    .await?;
-
-                start_ids
-            };
-
-            let (mut candidates, total_considered) = self
-                .index
-                .filter_search_candidates(&start_points, k, &mut self.scratch.best)
-                .await?;
-            self.scratch.best.drain_best(total_considered);
-
-            let computed_result_count = candidates.len();
-            self.computed_result.clear();
-            self.computed_result.append(&mut candidates);
-
-            self.next_result_index = 0;
-            if computed_result_count != self.search_param_l {
-                self.computed_result.truncate(computed_result_count);
-            }
-
-            let remaining_need = k - result.len();
-            let leftover = cmp::min(remaining_need, computed_result_count);
-            if leftover > 0 {
-                result.extend_from_slice(
-                    &self.computed_result
-                        [self.next_result_index..self.next_result_index + leftover],
-                );
-                self.next_result_index += leftover;
-            }
-
-            ANNResult::Ok(result)
-        }
-    }
 }
 
 #[cfg(test)]
