+++ diskann/src/graph/search/mod.rs
@@ -46,6 +46,9 @@ mod knn_search;
 mod multihop_search;
 mod range_search;
 
+mod paged;
+pub use paged::PagedSearch;
+
 pub mod record;
 pub(crate) mod scratch;
